This PR primarily does two things, chunks inserts of new fields into the `metabase_field` table and avoids updating fields when there has been no change. As a result, this significantly reduces the number of database calls made and the number of updates made.

On a test dataset I created that had a very high number of tables (2k) and columns (2M) I was able to average about 1.3ms per field. Because of the amount of I/O done previously, the time per field would fluctuate pretty significantly. That same dataset on the old code would average 1.2s per field over time. The first set of fields I saw closer to 12ms per row but it quickly increase to over 100ms and then after several hours it would creep up to over 1 second. I stopped the process after 4 hours and it inserted roughly 115k fields. The new code inserted all 2M in 54 minutes.